+++ src/Dap.Fabulous.Android/Decorator/NavigationPage.fs
@@ -1,5 +1,5 @@
 [<RequireQualifiedAccess>]
-module Dap.Fabulous.Android.Feature.Decorator.NavigationPage
+module Dap.Fabulous.Android.Decorator.NavigationPage
 
 open Xamarin.Forms
 open Xamarin.Forms.Platform.Android
@@ -9,6 +9,7 @@ open Dap.Prelude
 open Dap.Context
 open Dap.Platform
 
+open Dap.Fabulous.Controls
 open Dap.Fabulous.Decorator
 open Dap.Fabulous.Android
 
@@ -17,10 +18,9 @@ type AToolbar = Android.Support.V7.Widget.Toolbar
 // Xamarin.Forms.Platform.Android/AppCompat/NavigationPageRenderer.cs
 
 
-type Decorator (logging : ILogging) =
-    inherit EmptyContext (logging, SwitchCell.NativeDecoratorKind)
-    interface NavigationPage.INativeDecorator with
-        member this.SetBarActionColor (widget : NavigationPage) (color : Color) =
+type Decorator () =
+    interface INavigationPageDecorator with
+        member this.SetBarActionColor (widget : NavigationPage, color : Color) =
             // Can't change this easily, need to use Activity.SetTheme
             ()
         member this.UpdateBarStyle (widget : NavigationPage) =