+++ src/Dap.Fabulous.Android/Decorator/SwitchCell.fs
@@ -1,5 +1,5 @@
 [<RequireQualifiedAccess>]
-module Dap.Fabulous.Android.Feature.Decorator.SwitchCell
+module Dap.Fabulous.Android.Decorator.SwitchCell
 
 open Xamarin.Forms
 open Xamarin.Forms.Platform.Android
@@ -8,8 +8,9 @@ open Dap.Prelude
 open Dap.Context
 open Dap.Platform
 
+open Dap.Fabulous.Controls
 open Dap.Fabulous.Decorator
-open Xamarin.Forms
+open Dap.Fabulous.Android
 
 // Xamarin.Forms.Platform.Android/Cells/SwitchCellRenderer.cs
 
@@ -19,14 +20,14 @@ let getSwitchCellView (logger : ILogger) (cell : SwitchCell) : SwitchCellView op
         Util.getInstanceValue<SwitchCellRenderer, SwitchCellView> logger "_view" renderer
     )
 
-type Decorator (logging : ILogging) =
-    inherit EmptyContext (logging, SwitchCell.NativeDecoratorKind)
-    interface SwitchCell.INativeDecorator with
-        member this.SetTextColor (widget : SwitchCell) (color : Color) =
-            getSwitchCellView this widget
+type Decorator () =
+    let logger = (getLogging ()) .GetLogger "SwitchCellDecorator"
+    interface ISwitchCellDecorator with
+        member this.SetTextColor (widget : SwitchCell, color : Color) =
+            getSwitchCellView logger widget
             |> Option.iter (fun view ->
                 if view =? null then
-                    logError this "SetTextColor" "SwitchCellView_Not_Found" (widget, color)
+                    logError logger "SetTextColor" "SwitchCellView_Not_Found" (widget, color)
                 else
                     view.SetMainTextColor (color)
             )