+++ src/Dap.Fabulous.Mac/Decorator/SwitchCell.fs
@@ -1,5 +1,5 @@
 [<RequireQualifiedAccess>]
-module Dap.Fabulous.Mac.Feature.Decorator.SwitchCell
+module Dap.Fabulous.Mac.Decorator.SwitchCell
 
 open Foundation
 open CoreGraphics
@@ -12,13 +12,13 @@ open Dap.Prelude
 open Dap.Context
 open Dap.Platform
 
+open Dap.Fabulous.Controls
 open Dap.Fabulous.Decorator
 open Xamarin.Forms
 
-type Decorator (logging : ILogging) =
-    inherit EmptyContext (logging, SwitchCell.NativeDecoratorKind)
-    interface SwitchCell.INativeDecorator with
-        member this.SetTextColor (widget : SwitchCell) (color : Color) =
+type Decorator () =
+    interface ISwitchCellDecorator with
+        member this.SetTextColor (widget : SwitchCell, color : Color) =
             () //TODO
             (*
             Cell.getRealCell this widget