+++ src/Dap.Fabulous.UWP/Decorator/SwitchCellDecorator.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+using Xamarin.Forms;
+using Dap.Fabulous.Controls;
+
+namespace Dap.Fabulous.UWP {
+    public class SwitchCellDecorator : ISwitchCellDecorator {
+        public void SetTextColor(SwitchCell cell, Color color) {
+            //TODO
+        }
+    }
+}