+++ src/Dap.Fabulous.iOS/Decorator/SwitchCell.fs
@@ -1,5 +1,5 @@
 [<RequireQualifiedAccess>]
-module Dap.Fabulous.iOS.Feature.Decorator.SwitchCell
+module Dap.Fabulous.iOS.Decorator.SwitchCell
 
 open Foundation
 open CoreGraphics
@@ -15,10 +15,9 @@ open Dap.Platform
 open Dap.Fabulous.Decorator
 open Xamarin.Forms
 
-type Decorator (logging : ILogging) =
-    inherit EmptyContext (logging, SwitchCell.NativeDecoratorKind)
-    interface SwitchCell.INativeDecorator with
-        member this.SetTextColor (widget : SwitchCell) (color : Color) =
+type Decorator () =
+    interface ISwitchCellDecorator with
+        member this.SetTextColor (widget : SwitchCell, color : Color) =
             Cell.getRealCell this widget
             |> Option.iter (fun cell ->
                 cell.TextLabel.TextColor <- color.ToUIColor ()