+++ src/Dap.Fabulous.Controls/RELEASE_NOTES.md
@@ -1,2 +1,6 @@
 ## 0.6.0
+* TODO:RELEASE
+* Define native decorators here to provide UWP implementation in C#
+
+## 0.6.0
 * TextActionCell