+++ src/Dap.Fabulous/RELEASE_NOTES.md
@@ -1,4 +1,8 @@
 ## 0.6.4
+* TODO:RELEASE
+* Use INativeDecorator from Dap.Fabulous.Controls
+
+## 0.6.4
 * Create builders for easier gui creation with dsl
 * Support TextActionCell as custom cell